+++ dune/pdelab/backend/istl/vectorhelpers.hh
@@ -12,6 +12,7 @@
 
 #include <dune/pdelab/backend/istl/tags.hh>
 #include <dune/pdelab/backend/istl/descriptors.hh>
+#include <dune/pdelab/finiteelementmap/utility.hh>
 #include <dune/pdelab/ordering/orderingbase.hh>
 
 namespace Dune {
@@ -130,14 +131,17 @@ namespace Dune {
       struct vector_descriptor_helper
       {
         // export backend type, as the actual TMP is in the parent reduction functor
-        typedef typename Node::Traits::Backend type;
+        typedef Node type;
       };
 
       // descriptor for backends of leaf spaces collecting various information about
       // possible blocking structures
-      template<typename E, typename Backend>
+      template<typename E, typename GFS>
       struct leaf_vector_descriptor
       {
+
+        using Backend = typename GFS::Traits::Backend;
+        using FEM = typename GFS::Traits::FiniteElementMap;
 
         static_assert(Backend::Traits::block_type != Blocking::bcrs,
                       "Dynamically blocked leaf spaces are not supported by this backend.");
@@ -153,17 +157,37 @@ namespace Dune {
         static const bool support_cascaded_blocking =
           Backend::Traits::block_type == Blocking::none; // FIXME
 
+        // The cumulative block size is used by the algorithm to calculate total block
+        // size over several children for cascaded blocking. We try to extract this size
+        // from the finite element map (that works if the number of DOFs per entity is
+        // an identical constant for all types of entities that have DOFs attached), and
+        // if we fail we fall back to 1, which always works.
+        static const std::size_t detected_cumulative_block_size = Dune::Std::detected_or_t<
+          std::integral_constant<std::size_t,0>,
+          FiniteElementMapBlockSize,
+          FEM
+          >::value;
+
+        // Has the user selected an explicit block size? If yes, give priority to that.
+        static const std::size_t cumulative_block_size = Backend::Traits::block_size > 0
+          ? Backend::Traits::block_size
+          : detected_cumulative_block_size;
+
+        static constexpr bool have_valid_block_size = cumulative_block_size > 0;
+
+        static_assert(
+          Backend::Traits::block_size == 0 or (detected_cumulative_block_size % cumulative_block_size) == 0,
+          "The vector block size you specified is not compatible with the finite element map"
+          );
+
+        static_assert(
+          Backend::Traits::block_type != Blocking::fixed or have_valid_block_size,
+          "You requested static blocking, but we cannot extract a valid block size from the finite element map. Please specify the block size with the second template parameter of the vector backend."
+          );
+
         // The static block size of the associated vector
         static const std::size_t block_size =
-          Backend::Traits::block_type == Blocking::fixed
-          ? Backend::Traits::block_size
-          : 1;
-
-        // The cumulative block size is used by the algorithm to calculate total block
-        // size over several children for cascaded blocking. Right now, this will always be set to
-        // the block size passed in by the user, but it might also be possible to provide this
-        // information in the FiniteElementMap and provide automatic blocking detection.
-        static const std::size_t cumulative_block_size = Backend::Traits::block_size;
+          Backend::Traits::block_type == Blocking::fixed ? cumulative_block_size : 1;
 
         // The element type for the vector.
         typedef E element_type;
@@ -177,7 +201,7 @@ namespace Dune {
       template<typename E, typename Node, typename Tag>
         struct vector_descriptor_helper<E,Node,Tag, /* is LeafTag */ true>
       {
-        typedef leaf_vector_descriptor<E,typename Node::Traits::Backend> type;
+        typedef leaf_vector_descriptor<E,Node> type;
       };
 
       // the actual functor
@@ -220,6 +244,9 @@ namespace Dune {
             typename Sibling::vector_type,
             typename Child::vector_type
            >::value);
+
+        static constexpr bool have_valid_block_size =
+          Sibling::have_valid_block_size and Child::have_valid_block_size;
 
         // block size
         static const std::size_t block_size =
@@ -265,9 +292,13 @@ namespace Dune {
       };
 
       // Data part of child -> parent reduction descriptor
-      template<typename Child, typename Backend>
+      template<typename Child, typename GFS>
       struct parent_child_vector_descriptor_data
       {
+
+        using Backend = typename GFS::Traits::Backend;
+
+        static constexpr bool have_valid_block_size = Child::have_valid_block_size;
 
         // If all our have a common blocking structure, we can just
         // concatenate them without doing any blocking
@@ -280,11 +311,22 @@ namespace Dune {
           Child::support_cascaded_blocking &&
           Backend::Traits::block_type == Blocking::none;
 
+        // It is not allowed to specify a block size on an interior node
+        static_assert(
+          Backend::Traits::block_size == 0,
+          "You cannot specify a block size on interior nodes of the function space tree."
+          );
+
         // Throw an assertion if the user requests static blocking at this level,
         // but we cannot support it.
         static_assert((Backend::Traits::block_type != Blocking::fixed) ||
                       Child::support_cascaded_blocking,
                       "invalid blocking structure.");
+
+        static_assert(
+          Backend::Traits::block_type != Blocking::fixed or have_valid_block_size,
+          "You requested static blocking, but at least one leaf space has a finite element that does not support automatic block size extraction. Please specify the block size with the second template parameter of that space's vector backend."
+          );
 
         // If we block statically, we create bigger blocks, otherwise the
         // block size doesn't change.
@@ -362,15 +404,15 @@ namespace Dune {
       struct combine_vector_descriptor_parent
       {
 
-        template<typename Child, typename Backend>
+        template<typename Child, typename GFS>
         struct reduce
         {
 
           struct type
             : public parent_child_vector_descriptor<parent_child_vector_descriptor_data<
                                                       Child,
-                                                      Backend>,
-                                                    Backend::Traits::block_type
+                                                      GFS>,
+                                                    GFS::Traits::Backend::Traits::block_type
                                                     >
           {};
         };
@@ -388,6 +430,7 @@ namespace Dune {
       {};
 
     } // namespace ISTL
+
 
 #endif // DOXYGEN